Add files via upload
In statistics, exploratory data analysis is an approach of analyzing data sets to summarize their main characteristics, often using statistical graphics and other data visualization methods. Primarily EDA is for seeing what the data can tell us beyond the formal modeling or hypothesis testing task. EDA can be said to be an eye-opener to a dataset.